+++ NEWS.md
@@ -1,6 +1,7 @@
 # tsibble 0.9.1
 
 * Fixed bugs when `summarise()` with grouped factors. (#197)
+* Fixed `select()` when removing redundant key. (#196)
 
 # tsibble 0.9.0